+++ distributions/common/k8s/worker-deployment.yaml
@@ -16,7 +16,7 @@
 # -----------------------------------------------------------------------
 # Kubernetes deployment for workers. Workers each process one job, and
 # terminate themselves at the end of it. The "Always" restart policy in
-# this configuration is necessary to constantly refill the worker pool
+# this configuration is necessary to constantly refill the transfer pool
 # as workers complete jobs and terminate.
 #
 apiVersion: extensions/v1beta1
@@ -24,7 +24,7 @@ kind: Deployment
 metadata:
   labels:
     app: portability
-  name: portability-worker
+  name: portability-transfer
 spec:
   replicas: 3
   selector:
@@ -46,7 +46,7 @@ spec:
         secret:
           secretName: portability-service-account-creds
       containers:
-      - name: portability-worker
+      - name: portability-transfer
         image: IMAGE # Replaced by script
         imagePullPolicy: IfNotPresent
         volumeMounts:
