+++ distributions/demo-google-deployment/worker/build.gradle
@@ -25,7 +25,7 @@ repositories {
 }
 
 dependencies {
-    compile project(':portability-worker')
+    compile project(':portability-transfer')
 
     compile project(':extensions:data-transfer:portability-data-transfer-flickr')
     compile project(':extensions:data-transfer:portability-data-transfer-google')
@@ -36,7 +36,7 @@ dependencies {
 
 addCloudExtensionDependency(project)
 
-mainClassName = 'org.dataportabilityproject.worker.WorkerMain'
+mainClassName = 'org.dataportabilityproject.transfer.WorkerMain'
 
 copyResources()
