+++ distributions/demo-server/build.gradle
@@ -58,7 +58,7 @@ dependencies {
     compile project(':portability-bootstrap-vm')
 
     compile project(':portability-api')
-    compile project(':portability-worker')
+    compile project(':portability-transfer')
 
     compile project(':extensions:config:portability-config-yaml')
