+++ portability-spi-cloud/src/main/java/org/dataportabilityproject/spi/cloud/extension/CloudExtension.java
@@ -19,7 +19,7 @@ import org.dataportabilityproject.api.launcher.AbstractExtension;
 import org.dataportabilityproject.spi.cloud.storage.AppCredentialStore;
 import org.dataportabilityproject.spi.cloud.storage.JobStore;
 
-/** Cloud extensions implement this interface to be loaded in either a api or worker process. */
+/** Cloud extensions implement this interface to be loaded in either a api or transfer worker process. */
 public interface CloudExtension extends AbstractExtension {
 
   /**
