+++ portability-spi-transfer/src/main/java/org/dataportabilityproject/spi/transfer/extension/TransferExtension.java
@@ -20,7 +20,7 @@ import org.dataportabilityproject.spi.transfer.provider.Exporter;
 import org.dataportabilityproject.spi.transfer.provider.Importer;
 
 /**
- * Transfer extensions implement this contract to be loaded in a worker process.
+ * Transfer extensions implement this contract to be loaded in a transfer process.
  */
 public interface TransferExtension extends AbstractExtension {
