+++ portability-transfer/.gitignore
@@ -1,0 +1,5 @@
+*~
+target/**
+dependency-reduced-pom.xml
+# Resources are generated at each build and copied into src/main/resources
+src/main/resources/*
