+++ portability-transfer/build.gradle
@@ -37,7 +37,7 @@ jar {
         exclude '**/META-INF/*.RSA'
     }
     manifest {
-        attributes 'Main-Class': 'org.dataportabilityproject.worker.WorkerMain'
+        attributes 'Main-Class': 'org.dataportabilityproject.transfer.WorkerMain'
     }
 }
