+++ portability-transfer/src/main/java/org/dataportabilityproject/transfer/JobPoller.java
@@ -13,23 +13,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.worker;
+package org.dataportabilityproject.transfer;
 
 import com.google.inject.Inject;
 
-/** Polls and processes a {@code PortabilityJob}. */
-final class Worker {
-  private final JobPoller jobPoller;
-  private final JobProcessor jobProcessor;
+/**
+ * Polls a {@code PortabilityJob} for this transfer to process.
+ *
+ * <p>
+ *
+ * <p>Lightweight wrapper around an {@code AbstractScheduledService} so as to not expose its
+ * implementation details.
+ */
+final class JobPoller {
+  private final JobPollingService jobPollingService;
 
   @Inject
-  Worker(JobPoller jobPoller, JobProcessor jobProcessor) {
-    this.jobPoller = jobPoller;
-    this.jobProcessor = jobProcessor;
+  JobPoller(JobPollingService jobPollingService) {
+    this.jobPollingService = jobPollingService;
   }
 
-  void doWork() {
-    jobPoller.pollJob();
-    jobProcessor.processJob();
+  void pollJob() {
+    jobPollingService.startAsync();
+    jobPollingService.awaitTerminated();
   }
 }
