+++ portability-transfer/src/main/java/org/dataportabilityproject/transfer/WorkerModule.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.worker;
+package org.dataportabilityproject.transfer;
 
 import static com.google.common.collect.MoreCollectors.onlyElement;
